requirements.txt
python-jose
pydantic
fastapi
uvicorn
sqlalchemy
pandas
numpy
streamlit
requests
passlib[bcrypt]
python-multipart